% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest}
\alias{unnest}
\title{Unnest nested lists}
\usage{
unnest(
  x,
  spec = NULL,
  dedupe = FALSE,
  stack_atomic = FALSE,
  rep_to_max = FALSE
)
}
\arguments{
\item{x}{a nested list to unnest}

\item{spec}{spec to use for unnesting. See [`spec()`].}

\item{stack_atomic}{Whether atomic vectors should be stacked or not.}

\item{rep_to_max}{When the results of sibling nodes is joined, the shorter
(in terms o number of rows) components can be either recycled to the max
number of rows (`rep_to_max = TRUE`). This is R's standard recycling
behavior. Or, the results can be cross joined (aka form all combinations of
rows).}

\item{dedup}{whether to dedupe repeated elements. If TRUE, if a node is
visited for a second time and is not explicitly declared in the `spec` the
node is skipped. This is particularly useful with `group`ed specs.}
}
\description{
Unnest nested lists
}
